requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
black==20.8b1
boto3==1.16.30
botocore==1.19.30
certifi==2020.11.8
cffi==1.14.3
channels==3.0.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
dataclasses==0.6
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
django-cors-headers==3.5.0
django-extensions==3.1.0
django-filter==2.4.0
django-oauth2-provider==0.2.6.1
django-storages==1.10.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
jmespath==0.10.0
Markdown==3.3.3
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.1
whitenoise==5.2.0
zipp==3.4.0
zope.interface==5.2.0